Improvements in interrupts

+++ src/api/core/thread/atomic.d
@@ -28,11 +28,11 @@ version (Riscv64)
 
 unittest
 {
-    int v = 12;
+    size_t v = 12;
     auto res = cas(&v, 12, 22);
     assert(res);
     assert(v == 22);
 
-    int v1 = 12;
+    size_t v1 = 12;
     assert(!cas(&v1, 24, 22));
 }
